+++ include/EvoIndividual.hpp
@@ -14,5 +14,5 @@ public:
     std::vector<TransformXAllele> x_transformer_chromosome;
     std::vector<TransformYAllele> y_transformer_chromosome;
     std::vector<RobustAllele> robuster_chromosome;
-    float fitness;
+    double fitness;
 };
